crate: split backend implementations into modules
This helps to reduce #[cfg()] duplication.
